Reorg tensor collection
