+++ src/nn/tensor_collection/mod.rs
@@ -1,0 +1,5 @@
+mod collection;
+mod visitor;
+
+pub use collection::{TensorCollection, TensorOptions, TensorVisitor};
+pub use visitor::{RecursiveWalker, TensorContainer, TensorMut, TensorRef, VisitTensors};
